northbridge/amd/amdfam10: Use adequate size for HT speed limit field
Reviewed-on: http://review.coreboot.org/11587
Tested-by: build bot (Jenkins)
Tested-by: Raptor Engineering Automated Test Stand <noreply@raptorengineeringinc.com>

+++ src/northbridge/amd/amdfam10/amdfam10.h
@@ -1005,7 +1005,7 @@ struct nodes_info_t {
 } __attribute__((packed));
 
 struct ht_link_config {
-	uint8_t ht_speed_limit; // Speed in MHz; 0 for autodetect (default)
+	uint32_t ht_speed_limit; // Speed in MHz; 0 for autodetect (default)
 };
 
 /* be careful with the alignment of sysinfo, bacause sysinfo may be shared by coreboot_car and ramstage stage. and ramstage may be running at 64bit later.*/
